+++ src/main/java/gift/controller/OrderController.java
@@ -5,18 +5,14 @@ import gift.dto.betweenClient.ResponseDTO;
 import gift.dto.betweenClient.member.MemberDTO;
 import gift.dto.betweenClient.order.OrderRequestDTO;
 import gift.dto.betweenClient.order.OrderResponseDTO;
-import gift.entity.Option;
-import gift.service.KakaoTokenService;
-import gift.service.MemberService;
-import gift.service.OptionService;
 import gift.service.OrderService;
-import gift.service.WishListService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
@@ -27,20 +23,10 @@ import org.springframework.web.bind.annotation.RequestMapping;
 @Controller
 @RequestMapping("/api/orders")
 public class OrderController {
-
-    private final KakaoTokenService kakaoTokenService;
-    private final MemberService memberService;
-    private final OptionService optionService;
-    private final WishListService wishListService;
     private final OrderService orderService;
 
-    public OrderController(KakaoTokenService kakaoTokenService, MemberService memberService,
-            OptionService optionService, WishListService wishListService,
-            OrderService orderService) {
-        this.kakaoTokenService = kakaoTokenService;
-        this.memberService = memberService;
-        this.optionService = optionService;
-        this.wishListService = wishListService;
+    @Autowired
+    public OrderController(OrderService orderService) {
         this.orderService = orderService;
     }
 
@@ -54,19 +40,8 @@ public class OrderController {
                     content = @Content(mediaType = "application/json", schema = @Schema(implementation = ResponseDTO.class))),
             @ApiResponse(responseCode = "500", description = "서버에 의한 오류입니다.",
                     content = @Content(mediaType = "application/json", schema = @Schema(implementation = ResponseDTO.class)))})
-    public ResponseEntity<?> order(@Parameter(hidden = true) @LoginMember MemberDTO memberDTO, @RequestBody OrderRequestDTO orderRequestDTO) {
-        String accessToken = memberService.getMemberAccessToken(memberDTO.getEmail());
-
-        optionService.subtractOptionQuantity(orderRequestDTO.optionId(),
-                orderRequestDTO.quantity());
-
-        Option option = optionService.getOption(orderRequestDTO.optionId());
-
-
-        wishListService.removeWishListProduct(memberDTO, option.getProduct().getId());
-        kakaoTokenService.sendMsgToMe(accessToken, option, orderRequestDTO.message());
-        OrderResponseDTO orderResponseDTO = orderService.saveOrderHistory(orderRequestDTO);
-
-        return new ResponseEntity<>(orderResponseDTO, HttpStatus.CREATED);
+    public ResponseEntity<?> order(@Parameter(hidden = true) @LoginMember MemberDTO memberDTO, @RequestBody OrderRequestDTO orderRequestDTO){
+        return new ResponseEntity<>(orderService.order(memberDTO, orderRequestDTO), HttpStatus.CREATED);
     }
+
 }
