+++ src/main/java/gift/service/OrderService.java
@@ -1,41 +1,43 @@
 package gift.service;
 
+import gift.annotation.LoginMember;
+import gift.dto.betweenClient.member.MemberDTO;
 import gift.dto.betweenClient.order.OrderRequestDTO;
 import gift.dto.betweenClient.order.OrderResponseDTO;
 import gift.entity.Option;
-import gift.entity.OrderHistory;
-import gift.exception.BadRequestExceptions.BadRequestException;
-import gift.exception.InternalServerExceptions.InternalServerException;
-import gift.repository.OptionRepository;
-import gift.repository.OrderHistoryRepository;
-import java.time.LocalDateTime;
+import io.swagger.v3.oas.annotations.Parameter;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.bind.annotation.RequestBody;
 
 @Service
 public class OrderService {
 
-    private final OptionRepository optionRepository;
-    private final OrderHistoryRepository orderHistoryRepository;
+    private final KakaoTokenService kakaoTokenService;
+    private final MemberService memberService;
+    private final OptionService optionService;
+    private final WishListService wishListService;
+    private final OrderHistoryService orderHistoryService;
 
-    public OrderService(OptionRepository optionRepository,
-            OrderHistoryRepository orderHistoryRepository) {
-        this.optionRepository = optionRepository;
-        this.orderHistoryRepository = orderHistoryRepository;
+    public OrderService(KakaoTokenService kakaoTokenService, MemberService memberService,
+            OptionService optionService, WishListService wishListService,
+            OrderHistoryService orderHistoryService) {
+        this.kakaoTokenService = kakaoTokenService;
+        this.memberService = memberService;
+        this.optionService = optionService;
+        this.wishListService = wishListService;
+        this.orderHistoryService = orderHistoryService;
     }
 
     @Transactional
-    public OrderResponseDTO saveOrderHistory(OrderRequestDTO orderRequestDTO) {
-        try {
-            Option option = optionRepository.findById(orderRequestDTO.optionId())
-                    .orElseThrow(() -> new BadRequestException("해당 옵션 Id를 찾지 못했습니다."));
-            OrderHistory orderHistory = orderRequestDTO.convertToOrder(option, LocalDateTime.now());
+    public OrderResponseDTO order(@Parameter(hidden = true) @LoginMember MemberDTO memberDTO, @RequestBody OrderRequestDTO orderRequestDTO) {
+        String accessToken = memberService.getMemberAccessToken(memberDTO.getEmail());
+        optionService.subtractOptionQuantity(orderRequestDTO.optionId(),
+                orderRequestDTO.quantity());
 
-            return OrderResponseDTO.convertToDTO(orderHistoryRepository.save(orderHistory));
-        } catch (BadRequestException e) {
-            throw e;
-        } catch (Exception e) {
-            throw new InternalServerException(e.getMessage());
-        }
+        Option option = optionService.getOption(orderRequestDTO.optionId());
+        wishListService.removeWishListProduct(memberDTO, option.getProduct().getId());
+        kakaoTokenService.sendMsgToMe(accessToken, option, orderRequestDTO.message());
+        return orderHistoryService.saveOrderHistory(orderRequestDTO);
     }
 }
